utils/DataSetPacker/requirements.txt
numpy==1.26.3
scipy==1.11.4
pillow==10.2.0